decouple aenea server from platform dependent rpcs
